+++ spec/app/models/game.rb
@@ -1,8 +1,10 @@
 class Game
   include Mongoid::Document
+
   field :high_score, type: Integer, default: 500
   field :score, type: Integer, default: 0
   field :name
+
   belongs_to :person, index: true, validate: true
   belongs_to :parent, class_name: "Game", foreign_key: "parent-id"
   has_one :video, validate: false